+++ govcd/nsxt_ipsec_vpn_tunnel.go
@@ -12,15 +12,13 @@ import (
 	"github.com/vmware/go-vcloud-director/v2/util"
 )
 
-// NsxtIpSecVpnTunnel supports site-to-site policy-based IPsec VPN between an NSX-T Data Center Edge Gateway instance
-// and a remote site.
-// IPsec VPN offers site-to-site connectivity between an Edge Gateway and remote sites which also use NSX-T Data Center
-// or which have either third-party hardware routers or VPN gateways that support IPsec.
-// Policy-based IPsec VPN requires a VPN policy to be applied to packets to determine which traffic is to be protected
-// by IPsec before being passed through a VPN tunnel. This type of VPN is considered static because when a local network
-// topology and configuration change, the VPN policy settings must also be updated to accommodate the changes.
-// NSX-T Data Center Edge Gateways support split tunnel configuration, with IPsec traffic taking routing precedence.
-// VMware Cloud Director supports automatic route redistribution when you use IPsec VPN on an NSX-T edge gateway.
+// NsxtIpSecVpnTunnel offers site-to-site connectivity between an Edge Gateway and remote sites which also use NSX-T
+// Data Center or which have either third-party hardware routers or VPN gateways that support IPsec. Policy-based IPsec
+// VPN requires a VPN policy to be applied to packets to determine which traffic is to be protected by IPsec before
+// being passed through a VPN tunnel. This type of VPN is considered static because when a local network topology and
+// configuration change, the VPN policy settings must also be updated to accommodate the changes. NSX-T Data Center Edge
+// Gateways support split tunnel configuration, with IPsec traffic taking routing precedence. VMware Cloud Director
+// supports automatic route redistribution when you use IPsec VPN on an NSX-T edge gateway.
 type NsxtIpSecVpnTunnel struct {
 	NsxtIpSecVpn *types.NsxtIpSecVpnTunnel
 	client       *Client
@@ -28,10 +26,10 @@ type NsxtIpSecVpnTunnel struct {
 	edgeGatewayId string
 }
 
-// GetAllIpSecVpns returns all IPsec VPN configurations
-func (egw *NsxtEdgeGateway) GetAllIpSecVpns(queryParameters url.Values) ([]*NsxtIpSecVpnTunnel, error) {
+// GetAllIpSecVpnTunnels returns all IPsec VPN Tunnel configurations
+func (egw *NsxtEdgeGateway) GetAllIpSecVpnTunnels(queryParameters url.Values) ([]*NsxtIpSecVpnTunnel, error) {
 	client := egw.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpn
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnel
 	apiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
@@ -61,13 +59,14 @@ func (egw *NsxtEdgeGateway) GetAllIpSecVpns(queryParameters url.Values) ([]*Nsxt
 	return wrappedResponses, nil
 }
 
-func (egw *NsxtEdgeGateway) GetIpSecVpnById(id string) (*NsxtIpSecVpnTunnel, error) {
+//GetIpSecVpnTunnelById retrieves single IPsec VPN Tunnel by ID
+func (egw *NsxtEdgeGateway) GetIpSecVpnTunnelById(id string) (*NsxtIpSecVpnTunnel, error) {
 	if id == "" {
-		return nil, fmt.Errorf("canot find NSX-T IPsec VPN configuration without ID")
+		return nil, fmt.Errorf("canot find NSX-T IPsec VPN Tunnel configuration without ID")
 	}
 
 	client := egw.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpn
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnel
 	minimumApiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
@@ -92,14 +91,18 @@ func (egw *NsxtEdgeGateway) GetIpSecVpnById(id string) (*NsxtIpSecVpnTunnel, err
 	return returnObject, nil
 }
 
-func (egw *NsxtEdgeGateway) GetIpSecVpnByName(name string) (*NsxtIpSecVpnTunnel, error) {
+// GetIpSecVpnTunnelByName retrieves single IPsec VPN Tunnel by Name.
+//
+// Note. Name uniqueness is not enforced therefore it might exist a few IPsec VPN Tunnels with the same name.
+// An error will be returned in that case.
+func (egw *NsxtEdgeGateway) GetIpSecVpnTunnelByName(name string) (*NsxtIpSecVpnTunnel, error) {
 	if name == "" {
-		return nil, fmt.Errorf("canot find NSX-T IPsec VPN configuration without Name")
+		return nil, fmt.Errorf("canot find NSX-T IPsec VPN Tunnel configuration without Name")
 	}
 
-	allVpns, err := egw.GetAllIpSecVpns(nil)
+	allVpns, err := egw.GetAllIpSecVpnTunnels(nil)
 	if err != nil {
-		return nil, fmt.Errorf("error retrieving all NSX-T IPsec VPN configurations: %s", err)
+		return nil, fmt.Errorf("error retrieving all NSX-T IPsec VPN Tunnel configurations: %s", err)
 	}
 
 	var allResults []*NsxtIpSecVpnTunnel
@@ -111,20 +114,22 @@ func (egw *NsxtEdgeGateway) GetIpSecVpnByName(name string) (*NsxtIpSecVpnTunnel,
 	}
 
 	if len(allResults) > 1 {
-		return nil, fmt.Errorf("error - found %d NSX-T IPsec VPN configuratios with Name '%s'. Expected 1", len(allResults), name)
+		return nil, fmt.Errorf("error - found %d NSX-T IPsec VPN Tunnel configuratios with Name '%s'. Expected 1",
+			len(allResults), name)
 	}
 
 	if len(allResults) == 0 {
 		return nil, ErrorEntityNotFound
 	}
 
-	// Retrieving the object by ID, because only it includes Pre-shared Key
-	return egw.GetIpSecVpnById(allResults[0].NsxtIpSecVpn.ID)
+	// Retrieving again the object by ID, because only it includes Pre-shared Key
+	return egw.GetIpSecVpnTunnelById(allResults[0].NsxtIpSecVpn.ID)
 }
 
-func (egw *NsxtEdgeGateway) CreateIpSecVpn(ipSecVpnConfig *types.NsxtIpSecVpnTunnel) (*NsxtIpSecVpnTunnel, error) {
+// CreateIpSecVpnTunnel creates IPsec VPN Tunnel and returns it
+func (egw *NsxtEdgeGateway) CreateIpSecVpnTunnel(ipSecVpnConfig *types.NsxtIpSecVpnTunnel) (*NsxtIpSecVpnTunnel, error) {
 	client := egw.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpn
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnel
 	minimumApiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
@@ -137,45 +142,46 @@ func (egw *NsxtEdgeGateway) CreateIpSecVpn(ipSecVpnConfig *types.NsxtIpSecVpnTun
 
 	task, err := client.OpenApiPostItemAsync(minimumApiVersion, urlRef, nil, ipSecVpnConfig)
 	if err != nil {
-		return nil, fmt.Errorf("error creating NSX-T IPsec VPN configuration: %s", err)
+		return nil, fmt.Errorf("error creating NSX-T IPsec VPN Tunnel configuration: %s", err)
 	}
 
 	err = task.WaitTaskCompletion()
 	if err != nil {
-		return nil, fmt.Errorf("task failed while creating NSX-T IPsec VPN configuration: %s", err)
+		return nil, fmt.Errorf("task failed while creating NSX-T IPsec VPN Tunnel configuration: %s", err)
 	}
 
-	// filtering even by Name is not supported
-	allVpns, err := egw.GetAllIpSecVpns(nil)
+	// filtering even by Name is not supported on VCD side
+	allVpns, err := egw.GetAllIpSecVpnTunnels(nil)
 	if err != nil {
-		return nil, fmt.Errorf("error retrieving all NSX-T IPsec VPN configuration after creation: %s", err)
+		return nil, fmt.Errorf("error retrieving all NSX-T IPsec VPN Tunnel configuration after creation: %s", err)
 	}
 
 	for index, singleConfig := range allVpns {
 		if singleConfig.IsEqualTo(ipSecVpnConfig) {
 			// retrieve exact value by ID, because only this endpoint includes private key
-			ipSecVpn, err := egw.GetIpSecVpnById(allVpns[index].NsxtIpSecVpn.ID)
+			ipSecVpn, err := egw.GetIpSecVpnTunnelById(allVpns[index].NsxtIpSecVpn.ID)
 			if err != nil {
-				return nil, fmt.Errorf("error retrieving NSX-T IPsec VPN configuration: %s", err)
+				return nil, fmt.Errorf("error retrieving NSX-T IPsec VPN Tunnel configuration: %s", err)
 			}
 
 			return ipSecVpn, nil
 		}
 	}
 
-	return nil, fmt.Errorf("error finding NSX-T IPsec VPN configuration after creation: %s", ErrorEntityNotFound)
+	return nil, fmt.Errorf("error finding NSX-T IPsec VPN Tunnel configuration after creation: %s", ErrorEntityNotFound)
 }
 
+// Update updates NSX-T IPsec VPN Tunnel configuration with newly supplied data.
 func (ipSecVpn *NsxtIpSecVpnTunnel) Update(ipSecVpnConfig *types.NsxtIpSecVpnTunnel) (*NsxtIpSecVpnTunnel, error) {
 	client := ipSecVpn.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpn
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnel
 	apiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
 	}
 
 	if ipSecVpn.NsxtIpSecVpn.ID == "" {
-		return nil, fmt.Errorf("cannot update NSX-T IPsec VPN configuration without ID")
+		return nil, fmt.Errorf("cannot update NSX-T IPsec VPN Tunnel configuration without ID")
 	}
 
 	urlRef, err := client.OpenApiBuildEndpoint(fmt.Sprintf(endpoint, ipSecVpn.edgeGatewayId), ipSecVpn.NsxtIpSecVpn.ID)
@@ -191,23 +197,23 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) Update(ipSecVpnConfig *types.NsxtIpSecVpnTun
 
 	err = client.OpenApiPutItem(apiVersion, urlRef, nil, ipSecVpnConfig, returnObject.NsxtIpSecVpn)
 	if err != nil {
-		return nil, fmt.Errorf("error updating NSX-T IPsec VPN configuration: %s", err)
+		return nil, fmt.Errorf("error updating NSX-T IPsec VPN Tunnel configuration: %s", err)
 	}
 
 	return returnObject, nil
 }
 
-// Delete allows users to delete NSX-T Application Port Profile
+// Delete allows users to delete NSX-T IPsec VPN Tunnel
 func (ipSecVpn *NsxtIpSecVpnTunnel) Delete() error {
 	client := ipSecVpn.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpn
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnel
 	minimumApiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return err
 	}
 
 	if ipSecVpn.NsxtIpSecVpn.ID == "" {
-		return fmt.Errorf("cannot delete NSX-T IPsec VPN configuration without ID")
+		return fmt.Errorf("cannot delete NSX-T IPsec VPN Tunnel configuration without ID")
 	}
 
 	urlRef, err := ipSecVpn.client.OpenApiBuildEndpoint(fmt.Sprintf(endpoint, ipSecVpn.edgeGatewayId), ipSecVpn.NsxtIpSecVpn.ID)
@@ -216,9 +222,8 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) Delete() error {
 	}
 
 	err = ipSecVpn.client.OpenApiDeleteItem(minimumApiVersion, urlRef, nil)
-
 	if err != nil {
-		return fmt.Errorf("error deleting NSX-T IPsec VPN configuration: %s", err)
+		return fmt.Errorf("error deleting NSX-T IPsec VPN Tunnel configuration: %s", err)
 	}
 
 	return nil
@@ -226,17 +231,18 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) Delete() error {
 
 // GetStatus returns status of IPsec VPN Tunnel.
 //
-// Note. This is not being immediately populated and may appear after some time
+// Note. This is not being immediately populated and may appear after some time depending on
+// NsxtIpSecVpnTunnelSecurityProfile.DpdConfiguration
 func (ipSecVpn *NsxtIpSecVpnTunnel) GetStatus() (*types.NsxtIpSecVpnTunnelStatus, error) {
 	client := ipSecVpn.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnStatus
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnelStatus
 	minimumApiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
 	}
 
 	if ipSecVpn.NsxtIpSecVpn.ID == "" {
-		return nil, fmt.Errorf("cannot get NSX-T IPsec VPN status without ID")
+		return nil, fmt.Errorf("cannot get NSX-T IPsec VPN Tunnel status without ID")
 	}
 
 	urlRef, err := ipSecVpn.client.OpenApiBuildEndpoint(fmt.Sprintf(endpoint, ipSecVpn.edgeGatewayId, ipSecVpn.NsxtIpSecVpn.ID))
@@ -247,17 +253,18 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) GetStatus() (*types.NsxtIpSecVpnTunnelStatus
 	ipSecVpnTunnelStatus := &types.NsxtIpSecVpnTunnelStatus{}
 
 	err = ipSecVpn.client.OpenApiGetItem(minimumApiVersion, urlRef, nil, ipSecVpnTunnelStatus)
-
 	if err != nil {
-		return nil, fmt.Errorf("error deleting NSX-T IPsec VPN configuration: %s", err)
+		return nil, fmt.Errorf("error deleting NSX-T IPsec VPN Tunnel configuration: %s", err)
 	}
 
 	return ipSecVpnTunnelStatus, nil
 }
 
+// UpdateTunnelConnectionProperties allows user to customize IPsec VPN Tunnel Security Profile when the default one
+// does not fit requirements.
 func (ipSecVpn *NsxtIpSecVpnTunnel) UpdateTunnelConnectionProperties(ipSecVpnTunnelConnectionProperties *types.NsxtIpSecVpnTunnelSecurityProfile) (*types.NsxtIpSecVpnTunnelSecurityProfile, error) {
 	client := ipSecVpn.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnConnectionProperties
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnelConnectionProperties
 	minimumApiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
@@ -281,9 +288,10 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) UpdateTunnelConnectionProperties(ipSecVpnTun
 	return ipSecVpnTunnelProfile, nil
 }
 
+// GetTunnelConnectionProperties retrieves IPsec VPN Tunnel Security Profile
 func (ipSecVpn *NsxtIpSecVpnTunnel) GetTunnelConnectionProperties() (*types.NsxtIpSecVpnTunnelSecurityProfile, error) {
 	client := ipSecVpn.client
-	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnConnectionProperties
+	endpoint := types.OpenApiPathVersion1_0_0 + types.OpenApiEndpointIpSecVpnTunnelConnectionProperties
 	minimumApiVersion, err := client.checkOpenApiEndpointCompatibility(endpoint)
 	if err != nil {
 		return nil, err
@@ -307,8 +315,8 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) GetTunnelConnectionProperties() (*types.Nsxt
 	return ipSecVpnTunnelProfile, nil
 }
 
-// IsEqualTo helps to find NSX-T IPsec Configuration
-// Combination of LocalAddress and RemoteAddress has to be unique. This is a list of fields compared:
+// IsEqualTo helps to find NSX-T IPsec VPN Tunnel Configuration
+// Combination of LocalAddress and RemoteAddress has to be unique (enforced by API). This is a list of fields compared:
 // * Name
 // * Description
 // * Enabled
@@ -318,12 +326,12 @@ func (ipSecVpn *NsxtIpSecVpnTunnel) IsEqualTo(vpnConfig *types.NsxtIpSecVpnTunne
 	return ipSetVpnRulesEqual(ipSecVpn.NsxtIpSecVpn, vpnConfig)
 }
 
-// ipSetVpnRulesEqual performs comparison of two rules to ease lookup. This is a list of fields compared:
-//// * Name
-//// * Description
-//// * Enabled
-//// * LocalEndpoint.LocalAddress
-//// * RemoteEndpoint.RemoteAddress
+// ipSetVpnRulesEqual performs comparison of two NSX-T IPsec VPN Tunnels to ease lookup. This is a list of fields compared:
+// * Name
+// * Description
+// * Enabled
+// * LocalEndpoint.LocalAddress
+// * RemoteEndpoint.RemoteAddress
 func ipSetVpnRulesEqual(first, second *types.NsxtIpSecVpnTunnel) bool {
 	util.Logger.Println("comparing NSX-T IP Sev VPN configuration:")
 	util.Logger.Printf("%+v\n", first)